cloud_functions/update-daily-etf-prices/requirements.txt
# Function dependencies, for example:
# package>=version
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
db-dtypes==1.3.0
Deprecated==1.2.15
google-api-core==2.19.2
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-cloud-appengine-logging==1.5.0
google-cloud-audit-log==0.3.0
google-cloud-bigquery==3.25.0
google-cloud-bigquery-storage==2.26.0
google-cloud-core==2.4.1
google-cloud-logging==3.11.3
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.14.0
grpcio==1.66.1
grpcio-status==1.66.1
idna==3.10
importlib_metadata==8.5.0
numpy==1.26.4
oauthlib==3.2.2
opentelemetry-api==1.29.0
packaging==24.1
pandas==1.5.3
pandas-gbq==0.23.1
proto-plus==1.24.0
protobuf==5.28.1
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydata-google-auth==1.8.2
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==75.1.0
six==1.16.0
urllib3==2.2.3
wrapt==1.17.0
xlrd==2.0.1
zipp==3.21.0

# google-cloud-bigquery
# google-cloud-bigquery-storage
# google-cloud-logging
# pandas-gbq
# numpy==1.26.4
# pandas==1.5.3
# requests
# xlrd
# pyarrow
